requirements.txt
Flask
Flask-SQLAlchemy
waitress
Werkzeug
Pastes